// src/api/frame/DataType.ts
export enum DataType {
	NONE = '00',
	BOOL = '01',
	CHAR8 = '02',
	UCHAR8 = '03',
	INT16 = '04',
	UINT16 = '05',
	INT32 = '06',
	UINT32 = '07',
	INT64 = '08',
	UINT64 = '09',
	FLOAT32 = '0A',
	DOUBLE64 = '0B',
	BITFIELD = '0C',
	STRING = '0D',
	CONTAINER = '0E',
	TIMESTAMP = '0F',
	BYTEARRAY = '10',
	ERROR = 'FF',
	UNKNOWN = 'FE',
}